LPS-187086 SF, move to package that matches the interface it implements
